serving/app/house_price/requirements.txt
pandas==1.3.5
numpy==1.21.6
scikit-learn==1.0.2
python-dotenv==0.21.0
fastapi==0.88.0
uvicorn==0.16.0
locust==2.25.0
pyquery==2.0.0